+++ features/Context/FixturesContext.php
@@ -4,6 +4,9 @@ namespace Context;
 
 use Symfony\Component\HttpFoundation\File\File;
 use Doctrine\Common\Util\Inflector;
+use Behat\Gherkin\Node\TableNode;
+use Behat\MinkExtension\Context\RawMinkContext;
+use Behat\Gherkin\Node\PyStringNode;
 use Oro\Bundle\BatchBundle\Entity\JobInstance;
 use Oro\Bundle\UserBundle\Entity\Role;
 use Oro\Bundle\UserBundle\Entity\User;
@@ -20,9 +23,7 @@ use Pim\Bundle\CatalogBundle\Entity\Category;
 use Pim\Bundle\CatalogBundle\Model\ProductPrice;
 use Pim\Bundle\CatalogBundle\Model\Group;
 use Pim\Bundle\CatalogBundle\Model\Media;
-use Behat\Gherkin\Node\TableNode;
-use Behat\MinkExtension\Context\RawMinkContext;
-use Behat\Gherkin\Node\PyStringNode;
+use Pim\Bundle\FlexibleEntityBundle\Entity\Metric;
 
 /**
  * A context for creating entities
@@ -469,6 +470,28 @@ class FixturesContext extends RawMinkContext
                         }
 
                         $value->setData($optionValue);
+                    } elseif ($value->getAttribute()->getAttributeType() === $this->attributeTypes['metric']) {
+                        $metric = $value->getData();
+
+                        if (false === strpos($data['value'], ' ')) {
+                            throw new \InvalidArgumentException(
+                                sprintf(
+                                    'Metric value does not match expected format "<data> <unit>": %s',
+                                    $data['value']
+                                )
+                            );
+                        }
+                        list($data, $unit) = explode(' ', $data['value']);
+
+                        if (!$metric) {
+                            $metric = new Metric();
+                            $metric->setFamily($value->getAttribute()->getMetricFamily());
+                        }
+
+                        $metric->setData($data);
+                        $metric->setUnit($unit);
+
+                        $value->setMetric($metric);
                     } else {
                         $value->setData($data['value']);
                     }
@@ -987,6 +1010,38 @@ class FixturesContext extends RawMinkContext
     }
 
     /**
+     * @param TableNode $table
+     *
+     * @return null
+     * @Given /^the following CSV to import:$/
+     */
+    public function theFollowingCSVToImport(TableNode $table)
+    {
+        $delimiter = ';';
+
+        $data = $table->getRowsHash();
+        $columns = join($delimiter, array_keys($data));
+
+        $rows = array();
+        foreach ($data as $key => $values) {
+            foreach ($values as $index => $value) {
+                $value = in_array($value, array('yes', 'no')) ? (int) $value === 'yes' : $value;
+                $rows[$index][] = $value;
+            }
+        }
+        $rows = array_map(
+            function ($row) use ($delimiter) {
+                return join($delimiter, $row);
+            },
+            $rows
+        );
+
+        array_unshift($rows, $columns);
+
+        return $this->theFollowingFileToImport(new PyStringNode(join("\n", $rows)));
+    }
+
+    /**
      * @param string    $code
      * @param TableNode $table
      *
@@ -1083,6 +1138,18 @@ class FixturesContext extends RawMinkContext
             }
         )->toArray();
         assertEquals($this->listToArray($categoryCodes), $categories);
+    }
+
+    /**
+     * @Given /^I have configured channel "([^"]*)" with the following conversion options:$/
+     */
+    public function iHaveConfiguredChannelWithTheFollowingConversionOptions($channel, TableNode $conversionUnits)
+    {
+        $this
+            ->getEntityOrException('Channel', $channel)
+            ->setConversionUnits($conversionUnits->getRowsHash());
+
+        $this->flush();
     }
 
     /**
@@ -1289,7 +1356,7 @@ class FixturesContext extends RawMinkContext
             } else {
                 throw new \InvalidArgumentException(
                     sprintf(
-                        'Expecting metric family and default metric unit to be defined for attribute "%s"',
+                        'Expecting "metric family" and "default metric unit" columns to be defined for attribute "%s"',
                         $data['label']
                     )
                 );
@@ -1376,7 +1443,15 @@ class FixturesContext extends RawMinkContext
                         $value->addOption($option);
                     }
                 }
+                break;
 
+            case $this->attributeTypes['metric']:
+                list($data, $unit) = explode(' ', $data);
+                $metric = new Metric();
+                $metric->setFamily($attribute->getMetricFamily());
+                $metric->setData($data);
+                $metric->setUnit($unit);
+                $value->setData($metric);
                 break;
 
             default: