+++ src/Pim/Bundle/CatalogBundle/Form/Type/ConversionUnitsType.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Pim\Bundle\CatalogBundle\Form\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+
+/**
+ * Conversion units form type
+ *
+ * @author    Gildas Quemener <gildas@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class ConversionUnitsType extends AbstractType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        foreach ($options['units']['measures_config'] as $family => $config) {
+            $builder->add(
+                $family,
+                'choice',
+                array(
+                    'choices'     => array_combine(array_keys($config['units']), array_keys($config['units'])),
+                    'empty_value' => 'Do not convert',
+                    'required'    => false,
+                    'select2'     => true,
+                )
+            );
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setDefaultOptions(OptionsResolverInterface $resolver)
+    {
+        $resolver->setDefaults(
+            array(
+                'units' => array()
+            )
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'pim_catalog_conversion_units';
+    }
+}