+++ src/Pim/Bundle/FlexibleEntityBundle/Form/Type/MetricType.php
@@ -2,6 +2,10 @@
 
 namespace Pim\Bundle\FlexibleEntityBundle\Form\Type;
 
+use Symfony\Component\Form\FormInterface;
+
+use Symfony\Component\Form\FormView;
+
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\AbstractType;
@@ -20,11 +24,11 @@ class MetricType extends AbstractType
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $unitOptions['choices'] = $options['units'];
+        $unitOptions['choices'] = array_combine(array_keys($options['units']), array_keys($options['units']));
+        $unitOptions['select2'] = true;
         if ($options['default_unit']) {
             $unitOptions['preferred_choices'] = $options['default_unit'];
         }
-        $options['family'] = 'Weight';
 
         $builder
             ->add('id', 'hidden')